+++ internal/usecase/state_usecase.go
@@ -5,24 +5,28 @@ import (
 	"booktrading/internal/repository"
 )
 
+// StateUsecase определяет интерфейс для работы с состояниями книг
 type StateUsecase interface {
 	CreateState(dto *state.CreateStateDTO) (*state.State, error)
-	GetStateByID(id int64) (*state.State, error)
+	GetStateByID(id uint) (*state.State, error)
 	GetAllStates() ([]*state.State, error)
-	UpdateState(id int64, dto *state.UpdateStateDTO) (*state.State, error)
-	DeleteState(id int64) error
+	UpdateState(id uint, dto *state.UpdateStateDTO) (*state.State, error)
+	DeleteState(id uint) error
 }
 
+// stateUsecase реализует интерфейс StateUsecase
 type stateUsecase struct {
 	stateRepo repository.StateRepository
 }
 
+// NewStateUsecase создает новый экземпляр stateUsecase
 func NewStateUsecase(stateRepo repository.StateRepository) StateUsecase {
 	return &stateUsecase{
 		stateRepo: stateRepo,
 	}
 }
 
+// CreateState создает новое состояние
 func (u *stateUsecase) CreateState(dto *state.CreateStateDTO) (*state.State, error) {
 	s := &state.State{
 		Name: dto.Name,
@@ -35,15 +39,18 @@ func (u *stateUsecase) CreateState(dto *state.CreateStateDTO) (*state.State, err
 	return s, nil
 }
 
-func (u *stateUsecase) GetStateByID(id int64) (*state.State, error) {
+// GetStateByID получает состояние по ID
+func (u *stateUsecase) GetStateByID(id uint) (*state.State, error) {
 	return u.stateRepo.GetByID(id)
 }
 
+// GetAllStates получает список всех состояний
 func (u *stateUsecase) GetAllStates() ([]*state.State, error) {
 	return u.stateRepo.GetAll()
 }
 
-func (u *stateUsecase) UpdateState(id int64, dto *state.UpdateStateDTO) (*state.State, error) {
+// UpdateState обновляет существующее состояние
+func (u *stateUsecase) UpdateState(id uint, dto *state.UpdateStateDTO) (*state.State, error) {
 	s, err := u.stateRepo.GetByID(id)
 	if err != nil {
 		return nil, err
@@ -58,6 +65,7 @@ func (u *stateUsecase) UpdateState(id int64, dto *state.UpdateStateDTO) (*state.
 	return s, nil
 }
 
-func (u *stateUsecase) DeleteState(id int64) error {
+// DeleteState удаляет состояние по ID
+func (u *stateUsecase) DeleteState(id uint) error {
 	return u.stateRepo.Delete(id)
 }
